Python/requirements.txt
pydantic-settings
sqlalchemy
psycopg2-binary
google-auth
PyJWT
uvicorn
fastapi
requests